+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 ## Unreleased
+-   **breaking**: metric `jvm.nonheap.commited` renamed to `jvm.nonheap.committed` (issue #70)
 -   Fixed issue #69 (windows jdk compatibility) by disabling metrics not obtainable on windows
+
 ## 0.2.5
 ### Version updates
 -   `org.scalameta:scalafmt-core from 2.0.1 to 2.2.2`