+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # Changelog
 ## Unreleased
--   **breaking**: metric `jvm.nonheap.commited` renamed to `jvm.nonheap.committed` (issue #70)
--   Fixed issue #69 (windows jdk compatibility) by disabling metrics not obtainable on windows
+
+## 0.3.0
+-   **breaking**: metric `jvm.nonheap.commited` renamed to `jvm.nonheap.committed` (issue [issue #70](https://github.com/avast/datadog4s/issues/70))
+-   Fixed issue [issue #69](https://github.com/avast/datadog4s/issues/69) (windows jdk compatibility) by disabling metrics not obtainable on windows
 
 ## 0.2.5
 ### Version updates